Editor and More Cleanup -Editor added to solution -New project configurations made for building editor -Editor errors and warnings fixed except MFC multi-byte compatibility (will not run) -Redundant sound system reference removed -Betty volume fix -Volume initialization fix -'/' in includes replaced with '\'

+++ Source/Editor/Action.cpp
@@ -142,7 +142,7 @@ const char* ActionUndoMgr::GetUndoString()
 //***********************************************************************
 bool ActionUndoMgr::HaveRedo() const
 {
-	return m_CurrentPos + 1 != m_listUndoActions.Count();
+	return m_CurrentPos + 1 != (long)m_listUndoActions.Count();
 }
 
 //***********************************************************************
